Move some types into shared wgpu-types crate
