+++ Cargo.toml
@@ -3,4 +3,5 @@ members = [
 	"wgpu-core",
     "wgpu-native",
     "wgpu-remote",
+    "wgpu-types",
 ]
