+++ wgpu-core/Cargo.toml
@@ -17,6 +17,7 @@ license = "MPL-2.0"
 [features]
 default = []
 metal-auto-capture = ["gfx-backend-metal/auto-capture"]
+serde1 = ["wgt/serde"]
 #NOTE: glutin feature is not stable, use at your own risk
 #glutin = ["gfx-backend-gl/glutin"]
 
@@ -35,6 +36,11 @@ rendy-descriptor = "0.5"
 serde = { version = "1.0", features = ["serde_derive"], optional = true }
 smallvec = "1.0"
 vec_map = "0.8"
+
+[dependencies.wgt]
+path = "../wgpu-types"
+package = "wgpu-types"
+version = "0.1"
 
 [target.'cfg(any(target_os = "ios", target_os = "macos"))'.dependencies]
 gfx-backend-metal = { version = "0.4" }
