+++ wgpu-core/src/resource.rs
@@ -5,55 +5,19 @@
 use crate::{
     id::{DeviceId, SwapChainId, TextureId},
     track::DUMMY_SELECTOR,
-    BufferAddress,
     Extent3d,
     LifeGuard,
     RefCount,
     Stored,
 };
 
+use wgt::{BufferAddress, BufferUsage, CompareFunction, TextureFormat};
 use hal;
 use rendy_memory::MemoryBlock;
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
 
 use std::{borrow::Borrow, fmt};
-
-bitflags::bitflags! {
-    #[repr(transparent)]
-    #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-    pub struct BufferUsage: u32 {
-        const MAP_READ = 1;
-        const MAP_WRITE = 2;
-        const COPY_SRC = 4;
-        const COPY_DST = 8;
-        const INDEX = 16;
-        const VERTEX = 32;
-        const UNIFORM = 64;
-        const STORAGE = 128;
-        const INDIRECT = 256;
-        const STORAGE_READ = 512;
-        const NONE = 0;
-        /// The combination of all read-only usages.
-        const READ_ALL = Self::MAP_READ.bits | Self::COPY_SRC.bits |
-            Self::INDEX.bits | Self::VERTEX.bits | Self::UNIFORM.bits |
-            Self::STORAGE_READ.bits | Self::INDIRECT.bits;
-        /// The combination of all write-only and read-write usages.
-        const WRITE_ALL = Self::MAP_WRITE.bits | Self::COPY_DST.bits | Self::STORAGE.bits;
-        /// The combination of all usages that the are guaranteed to be be ordered by the hardware.
-        /// If a usage is not ordered, then even if it doesn't change between draw calls, there
-        /// still need to be pipeline barriers inserted for synchronization.
-        const ORDERED = Self::READ_ALL.bits;
-    }
-}
-
-#[repr(C)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-#[derive(Clone, Debug)]
-pub struct BufferDescriptor {
-    pub size: BufferAddress,
-    pub usage: BufferUsage,
-}
 
 #[repr(C)]
 #[derive(Debug)]
@@ -139,69 +103,6 @@ pub enum TextureDimension {
     D3,
 }
 
-#[repr(C)]
-#[derive(Copy, Clone, Debug, Hash, Eq, PartialEq)]
-pub enum TextureFormat {
-    // Normal 8 bit formats
-    R8Unorm = 0,
-    R8Snorm = 1,
-    R8Uint = 2,
-    R8Sint = 3,
-
-    // Normal 16 bit formats
-    R16Unorm = 4,
-    R16Snorm = 5,
-    R16Uint = 6,
-    R16Sint = 7,
-    R16Float = 8,
-
-    Rg8Unorm = 9,
-    Rg8Snorm = 10,
-    Rg8Uint = 11,
-    Rg8Sint = 12,
-
-    // Normal 32 bit formats
-    R32Uint = 13,
-    R32Sint = 14,
-    R32Float = 15,
-    Rg16Unorm = 16,
-    Rg16Snorm = 17,
-    Rg16Uint = 18,
-    Rg16Sint = 19,
-    Rg16Float = 20,
-    Rgba8Unorm = 21,
-    Rgba8UnormSrgb = 22,
-    Rgba8Snorm = 23,
-    Rgba8Uint = 24,
-    Rgba8Sint = 25,
-    Bgra8Unorm = 26,
-    Bgra8UnormSrgb = 27,
-
-    // Packed 32 bit formats
-    Rgb10a2Unorm = 28,
-    Rg11b10Float = 29,
-
-    // Normal 64 bit formats
-    Rg32Uint = 30,
-    Rg32Sint = 31,
-    Rg32Float = 32,
-    Rgba16Unorm = 33,
-    Rgba16Snorm = 34,
-    Rgba16Uint = 35,
-    Rgba16Sint = 36,
-    Rgba16Float = 37,
-
-    // Normal 128 bit formats
-    Rgba32Uint = 38,
-    Rgba32Sint = 39,
-    Rgba32Float = 40,
-
-    // Depth and stencil formats
-    Depth32Float = 41,
-    Depth24Plus = 42,
-    Depth24PlusStencil8 = 43,
-}
-
 bitflags::bitflags! {
     #[repr(transparent)]
     pub struct TextureUsage: u32 {
@@ -274,22 +175,10 @@ impl Default for TextureAspect {
 }
 
 #[repr(C)]
-#[derive(Copy, Clone, Debug, Hash, Eq, PartialEq)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub enum TextureViewDimension {
-    D1,
-    D2,
-    D2Array,
-    Cube,
-    CubeArray,
-    D3,
-}
-
-#[repr(C)]
 #[derive(Debug)]
 pub struct TextureViewDescriptor {
     pub format: TextureFormat,
-    pub dimension: TextureViewDimension,
+    pub dimension: wgt::TextureViewDimension,
     pub aspect: TextureAspect,
     pub base_mip_level: u32,
     pub level_count: u32,
@@ -356,28 +245,6 @@ pub enum FilterMode {
 impl Default for FilterMode {
     fn default() -> Self {
         FilterMode::Nearest
-    }
-}
-
-#[repr(C)]
-#[derive(Copy, Clone, Debug, Hash, Eq, PartialEq)]
-pub enum CompareFunction {
-    Never = 0,
-    Less = 1,
-    Equal = 2,
-    LessEqual = 3,
-    Greater = 4,
-    NotEqual = 5,
-    GreaterEqual = 6,
-    Always = 7,
-}
-
-impl CompareFunction {
-    pub fn is_trivial(self) -> bool {
-        match self {
-            CompareFunction::Never | CompareFunction::Always => true,
-            _ => false,
-        }
     }
 }
