+++ wgpu-remote/cbindgen.toml
@@ -24,9 +24,9 @@ exclude = ["BufferMapResult"]
 
 [parse]
 parse_deps = true
-include = ["wgpu-core"]
+include = ["wgpu-core", "wgpu-types"]
 
-extra_bindings = ["wgpu-core"]
+extra_bindings = ["wgpu-core", "wgpu-types"]
 
 [fn]
 prefix = "WGPU_INLINE"
