+++ wgpu-remote/src/server.rs
@@ -43,7 +43,7 @@ pub extern "C" fn wgpu_server_poll_all_devices(global: &Global, force_wait: bool
 #[no_mangle]
 pub unsafe extern "C" fn wgpu_server_instance_request_adapter(
     global: &Global,
-    desc: &core::instance::RequestAdapterOptions,
+    desc: &wgt::RequestAdapterOptions,
     ids: *const id::AdapterId,
     id_length: usize,
 ) -> i8 {
@@ -61,7 +61,7 @@ pub unsafe extern "C" fn wgpu_server_instance_request_adapter(
 pub extern "C" fn wgpu_server_adapter_request_device(
     global: &Global,
     self_id: id::AdapterId,
-    desc: &core::instance::DeviceDescriptor,
+    desc: &wgt::DeviceDescriptor,
     new_id: id::DeviceId,
 ) {
     gfx_select!(self_id => global.adapter_request_device(self_id, desc, new_id));
@@ -84,7 +84,7 @@ pub extern "C" fn wgpu_server_device_destroy(global: &Global, self_id: id::Devic
 pub extern "C" fn wgpu_server_device_create_buffer(
     global: &Global,
     self_id: id::DeviceId,
-    desc: &core::resource::BufferDescriptor,
+    desc: &wgt::BufferDescriptor,
     new_id: id::BufferId,
 ) {
     gfx_select!(self_id => global.device_create_buffer(self_id, desc, new_id));
@@ -99,9 +99,9 @@ pub unsafe extern "C" fn wgpu_server_device_set_buffer_sub_data(
     global: &Global,
     self_id: id::DeviceId,
     buffer_id: id::BufferId,
-    offset: core::BufferAddress,
+    offset: wgt::BufferAddress,
     data: *const u8,
-    size: core::BufferAddress,
+    size: wgt::BufferAddress,
 ) {
     let slice = slice::from_raw_parts(data, size as usize);
     gfx_select!(self_id => global.device_set_buffer_sub_data(self_id, buffer_id, offset, slice));
@@ -115,8 +115,8 @@ pub unsafe extern "C" fn wgpu_server_device_set_buffer_sub_data(
 pub extern "C" fn wgpu_server_buffer_map_read(
     global: &Global,
     buffer_id: id::BufferId,
-    start: core::BufferAddress,
-    size: core::BufferAddress,
+    start: wgt::BufferAddress,
+    size: wgt::BufferAddress,
     callback: core::device::BufferMapReadCallback,
     userdata: *mut u8,
 ) {
@@ -127,7 +127,7 @@ pub extern "C" fn wgpu_server_buffer_map_read(
     );
     gfx_select!(buffer_id => global.buffer_map_async(
         buffer_id,
-        core::resource::BufferUsage::MAP_READ,
+        wgt::BufferUsage::MAP_READ,
         start .. start + size,
         operation
     ));
@@ -182,10 +182,10 @@ pub unsafe extern "C" fn wgpu_server_encoder_copy_buffer_to_buffer(
     global: &Global,
     self_id: id::CommandEncoderId,
     source_id: id::BufferId,
-    source_offset: core::BufferAddress,
+    source_offset: wgt::BufferAddress,
     destination_id: id::BufferId,
-    destination_offset: core::BufferAddress,
-    size: core::BufferAddress,
+    destination_offset: wgt::BufferAddress,
+    size: wgt::BufferAddress,
 ) {
     gfx_select!(self_id => global.command_encoder_copy_buffer_to_buffer(self_id, source_id, source_offset, destination_id, destination_offset, size));
 }
