+++ wgpu-types/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "wgpu-types"
+version = "0.1.0"
+authors = [
+	"Dzmitry Malyshau <kvark@mozilla.com>",
+	"Joshua Groves <josh@joshgroves.com>",
+]
+edition = "2018"
+description = "WebGPU types"
+homepage = "https://github.com/gfx-rs/wgpu"
+repository = "https://github.com/gfx-rs/wgpu"
+keywords = ["graphics"]
+license = "MPL-2.0"
+
+[lib]
+
+[dependencies]
+bitflags = "1.0"
+serde = { version = "1.0", features = ["serde_derive"], optional = true }
